+++ src/util/memmove.h
@@ -1,8 +1,7 @@
 #pragma once
 
 // Copy n bytes from src to dest. The src and dest regions may overlap.
-static void* memmove(void* dest, const void* src,
-                    unsigned int n) {
+static void* memmove(void* dest, const void* src, size_t n) {
   if (dest < src) {
     char* o = dest;
     char* const end = o + n;
