+++ src/util/strncmp.h
@@ -5,7 +5,7 @@
 //   <0: l.substr(0, n) < r.substr(0, n)
 //    0: l.substr(0, n) == r.substr(0, n)
 //   >0: l.substr(0, n) > r.substr(0, n)
-static int strncmp(const char* l, const char* r, unsigned n) {
+static int strncmp(const char* l, const char* r, size_t n) {
   unsigned char lc = 0, rc = 0;
   for (unsigned i = 0; i < n && lc == rc; i++) {
     lc = *l++, rc = *r++;
